Final changes and tweaks
We can change the text later but we will need another script for
that. I'll get around to it if it's needed.

I removed some text from the README to make it simpler (there's
already a guide inside every ticket). The template was improved a bit,
with more text added. A lot of text is bad, but making it unclear what
to do is probably worse… so I added some bold text to help with
skimming¹.

¹ – https://communitywiki.org/wiki/WriteForSkimming